+++ brainscore_language/model_helpers/huggingface.py
@@ -15,8 +15,8 @@ from transformers.modeling_outputs import CausalLMOutput
 
 from brainio.assemblies import DataAssembly, NeuroidAssembly, BehavioralAssembly
 from brainscore_language.artificial_subject import ArtificialSubject
+from brainscore_language.model_helpers.preprocessing import prepare_context
 from brainscore_language.utils import fullname
-from brainscore_language.utils.preprocessing import prepare_context
 
 
 class HuggingfaceSubject(ArtificialSubject):
